Add Terms and Materials
These two models are closely interrelated, so it made sense to combine
this work. There's some interesting stuff going on with id validation,
since neither of these two models has "normal" id structure, but it
seems to be working well. Maybe move that logic to the model?

The include definitions work, but the server doesn't accept an include
param via the query string. We don't often need to specify includes in
a data service request, but it's something to look into.