Undo aggressive stripping of defaults
It seems that Laravel does not have a mechanism for "fallback" to its
default configs if the files in the `configs` folder are deleted. This
is unfortunate, since most of this configuration is better done through
envars in `.env`. We should add a `DefaultConfigServiceProvider` in the
foundation to provide fallbacks if these files are omitted.

It seems that some default providers cannot be disabled without major
work-arounds, so I'm leaving them in for now. Specifically, I'd like to
have a way to get rid of Blade and Views, since our data services don't
use these features.